feat: added dog breeds to dog table

+++ database/migrations/2025_02_24_000001_create_dogs_table.php
@@ -9,13 +9,14 @@ return new class extends Migration
     /**
      * Run the migrations.
      */
-    public function up()
+    public function up(): void
     {
         Schema::create('dogs', function (Blueprint $table) {
             $table->id();
             $table->string('name');
             $table->string('nickname');
             $table->date('birthday');
+            $table->string('breed');
             $table->enum('gender', ['male', 'female']);
             $table->string('pedigree_url')->nullable();
             $table->foreignId('group_id')->constrained('dog_group_types');
@@ -28,31 +29,12 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::table('dogs', function (Blueprint $table) {
-            $table->dropForeign(['group_id']);
-        });
+        Schema::disableForeignKeyConstraints();
 
-        Schema::table('dog_owners', function (Blueprint $table) {
-            $table->dropForeign(['dog_id']);
-        });
-        Schema::table('medical_examinations', function (Blueprint $table) {
-            $table->dropForeign(['dog_id']);
-        });
-        Schema::table('dog_titles', function (Blueprint $table) {
-            $table->dropForeign(['dog_id']);
-        });
+        if (Schema::hasTable('dogs')) {
+            Schema::dropIfExists('dogs');
+        }
 
-        Schema::table('litters', function (Blueprint $table) {
-            $table->dropForeign(['mother_id']);
-        });
-        Schema::table('litters', function (Blueprint $table) {
-            $table->dropForeign(['father_id']);
-        });
-
-        Schema::table('litter_puppies', function (Blueprint $table) {
-            $table->dropForeign(['dog_id']);
-        });
-
-        Schema::dropIfExists('dogs');
+        Schema::enableForeignKeyConstraints();
     }
 };
